gtkui: mac compile fix

+++ plugins/gtkui/gtkui.c
@@ -164,10 +164,10 @@ gtkui_dispatch_on_main (void (^block) (void)) {
 static void
 format_timestr (char *buf, int sz, float time) {
     time = roundf (time);
-    int daystotal = (int64_t)time / (3600 * 24);
-    int hourtotal = ((int64_t)time / 3600) % 24;
-    int mintotal = ((int64_t)time / 60) % 60;
-    int sectotal = ((int64_t)time) % 60;
+    int daystotal = (int)(int64_t)time / (3600 * 24);
+    int hourtotal = (int)((int64_t)time / 3600) % 24;
+    int mintotal = (int)((int64_t)time / 60) % 60;
+    int sectotal = (int)((int64_t)time) % 60;
 
     if (daystotal == 0) {
         snprintf (buf, sz, "%d:%02d:%02d", hourtotal, mintotal, sectotal);
